Move AccountKey to be defined in a common library
AccountKey was used by both ACMECLI and ACMEKestrel, however it was only defined in ACMECLI.  The was causing ACMEKestrel to not compile.  To remedy this AccountKey was moved into Examples.Common.PKI.
